Moves usePrevious to utils and removes index file
