+++ packages/palette/src/elements/index.tsx
@@ -16,7 +16,7 @@ export * from "./Input"
 export * from "./Join"
 export * from "./Menu"
 export * from "./Message"
-export * from "./Modal"
+export * from "./Modal/Modal"
 export * from "./Pagination"
 export * from "./LabeledRange"
 export * from "./LineChart"
