+++ packages/palette/src/helpers/index.ts
@@ -2,3 +2,4 @@ export * from "./color"
 export * from "./injectGlobalStyles"
 export * from "./media"
 export * from "./space"
+export * from "./usePrevious"
