+++ packages/palette/src/helpers/usePrevious.tsx
@@ -1,0 +1,9 @@
+import { useEffect, useRef } from "react"
+
+export const usePrevious = value => {
+  const ref = useRef(null)
+  useEffect(() => {
+    ref.current = value
+  }, [value])
+  return ref.current
+}
